pyproject.toml
[project]
name = "conductor-cli"
version = "0.1.0"
description = "A CLI tool for defining and running multi-agent workflows with the GitHub Copilot SDK"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Conductor Team" }
]
keywords = [
    "cli",
    "copilot",
    "multi-agent",
    "workflow",
    "orchestration",
    "ai",
    "llm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "ruamel.yaml>=0.18.0",
    "jinja2>=3.1.0",
    "simpleeval>=1.0.0",
    "github-copilot-sdk>=0.1.0",
    "anthropic>=0.77.0,<1.0.0",
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/microsoft/conductor"
Documentation = "https://github.com/microsoft/conductor#readme"
Repository = "https://github.com/microsoft/conductor"
Issues = "https://github.com/microsoft/conductor/issues"
Changelog = "https://github.com/microsoft/conductor/releases"

[project.scripts]
conductor = "conductor.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/conductor"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "performance: marks tests as performance tests (deselect with '-m \"not performance\"')",
    "real_api: marks tests that make real API calls to external services (deselect with '-m \"not real_api\"')",
]
